refactor(quiz): rename QuizzesPage to ProfileQuizzesPage
- Update main.tsx route import from QuizzesPage → ProfileQuizzesPage
- Rename pages/QuizzesPage directory to pages/ProfileQuizzesPage
- Rename component QuizzesPage → ProfileQuizzesPage and update default export
- Rename QuizzesPageUI → ProfileQuizzesPageUI, including props and UISearchParams interfaces
- Update stories and tests to reference ProfileQuizzesPageUI
- Remove old QuizzesPage index and component files
