+++ packages/quiz/src/pages/ProfileQuizzesPage/components/ProfileQuizzesPageUI/index.ts
@@ -1,0 +1,5 @@
+export type {
+  ProfileQuizzesPageUIProps,
+  ProfileQuizzesPageUISearchParams,
+} from './ProfileQuizzesPageUI'
+export { default } from './ProfileQuizzesPageUI'
