+++ packages/quiz/src/pages/ProfileQuizzesPage/components/index.ts
@@ -1,0 +1,5 @@
+export type {
+  ProfileQuizzesPageUIProps,
+  ProfileQuizzesPageUISearchParams,
+} from './ProfileQuizzesPageUI'
+export { default as ProfileQuizzesPageUI } from './ProfileQuizzesPageUI'
