+++ packages/quiz/src/pages/ProfileQuizzesPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProfileQuizzesPage'
