--- packages/quiz/src/pages/QuizzesPage/components/index.ts
@@ -1,5 +1,0 @@
-export type {
-  QuizzesPageUIProps,
-  QuizzesPageUISearchParams,
-} from './QuizzesPageUI'
-export { default as QuizzesPageUI } from './QuizzesPageUI'
