--- packages/quiz/src/pages/QuizzesPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './QuizzesPage'
