+++ packages/quiz/src/pages/index.ts
@@ -10,4 +10,4 @@ export { default as LoginPage } from './LoginPage'
 export { default as ProfilePage } from './ProfilePage'
 export { default as QuizCreatorPage } from './QuizCreatorPage'
 export { default as QuizDetailsPage } from './QuizDetailsPage'
-export { default as QuizzesPage } from './QuizzesPage'
+export { default as ProfileQuizzesPage } from './ProfileQuizzesPage'
